refactor(autoreview): Move ExtensionPoint selector wiht the other seletors
